fix(fas): fd leak, verdict persistence, silence-gate integrity, DSP math (bug-hunt batch)
- fingerprint: close the mkstemp fd -- every scoring tick leaked one
  OS handle AND stranded a ~128 KB temp WAV (the open handle made the
  unlink fail silently on Windows). Long FAS sessions exhausted the
  fd limit. A one-shot startup sweep removes orphans left by prior
  builds.
- engine: hold/unhold and codec re-INVITEs no longer wipe per-call
  scoring state. Media re-attach now preserves _CallScoreState
  (committed verdict, sticky evidence, AASIST history, tick schedule)
  and only a true call end untracks -- a single unhold used to reset
  a locked CONFIRMED_FAS badge back to ANALYZING.
- worker: silence-gated ticks no longer feed the stale AASIST median
  into synthesise() -- an early spoof read used to keep firing on
  minutes of dead air, escalating to PROBABLE_FAS on windows the
  model never saw (exactly what the silence gate exists to prevent).
- features: Goertzel bin index is int() per the standard formula --
  every tone was evaluated half a bin (+5 Hz at 100 ms frames) off
  target, attenuating on-spec PSTN tones ~36%. On-frequency pure
  tones now score ~0.30 vs ~0.19; verified no rule-gate regressions
  (the 0.15 ringback gate was already satisfied pre-fix).
- router: a ring buffer holding exactly RING_SAMPLES no longer
  returns an empty snapshot (regression test added).
- replay: non-16-bit WAVs now raise a loud, file-named error instead
  of decoding shuffled bytes into a confident wrong verdict in the
  offline evidence report.

+++ python-app/src/noc_beam/audio/fas_features.py
@@ -64,7 +64,17 @@ def goertzel_magnitude(samples: np.ndarray, sample_rate: int, target_hz: float) 
         return 0.0
     x = samples.astype(np.float64) / 32768.0
     n = x.size
-    k = 0.5 + (n * target_hz) / sample_rate
+    # Standard Goertzel: the bin index k must be an INTEGER (the algorithm
+    # only evaluates the DFT at integer bins). The old code omitted the
+    # int(), so k landed half a bin high -- at a 100 ms / 16 kHz frame that
+    # is +5 Hz of detuning, evaluating every "on-frequency" tone off-target
+    # and attenuating a true match by up to ~36%. int(0.5 + ...) rounds the
+    # target frequency to the nearest bin, as intended.
+    # NOTE: tone thresholds elsewhere (fas_rules.py's 0.35 tone_score gate;
+    # "clean tones normalise near 0.20") were empirically tuned with the
+    # off-by-half-bin attenuation baked in, so post-fix tone/ringback scores
+    # read slightly HIGHER for on-frequency tones. See the FAS corpus gate.
+    k = int(0.5 + (n * target_hz) / sample_rate)
     omega = 2.0 * math.pi * k / n
     coeff = 2.0 * math.cos(omega)
     s_prev = 0.0
